PDFBOX-4892: Sonar fix / revert
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1886753 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/cos/COSNumber.java
@@ -62,7 +62,7 @@ public abstract class COSNumber extends COSBase
             char digit = number.charAt(0);
             if ('0' <= digit && digit <= '9') 
             {
-                return COSInteger.get(digit - '0');
+                return COSInteger.get((long) digit - '0');
             } 
             else if (digit == '-' || digit == '.') 
             {
